Fix float handling in poco FPEnvironment_SUN.cpp
The SunOS FP environment backend fails `-Wdouble-promotion`:

- `isInfiniteImpl(float)` passed a float to `fpclass`, which takes a
  double. Make the exact widening conversion explicit, as the
  `long double` overload already does.
- `copySignImpl(float, float)` promoted both arguments to call the
  double `copysign` and cast the result back. Use `copysignf`.
- `copySignImpl(double, double)` truncated the `copysign` result
  through a stray `(float)` cast before returning it as double,
  silently losing precision. Drop the cast.

+++ base/poco/Foundation/src/FPEnvironment_SUN.cpp
@@ -50,7 +50,7 @@ FPEnvironmentImpl& FPEnvironmentImpl::operator = (const FPEnvironmentImpl& env)
 
 bool FPEnvironmentImpl::isInfiniteImpl(float value)
 {
-	int cls = fpclass(value);
+	int cls = fpclass(static_cast<double>(value));
 	return cls == FP_PINF || cls == FP_NINF;
 }
 
@@ -89,13 +89,13 @@ bool FPEnvironmentImpl::isNaNImpl(long double value)
 
 float FPEnvironmentImpl::copySignImpl(float target, float source)
 {
-	return (float) copysign(target, source);
+	return copysignf(target, source);
 }
 
 
 double FPEnvironmentImpl::copySignImpl(double target, double source)
 {
-	return (float) copysign(target, source);
+	return copysign(target, source);
 }
